add dtype

+++ tests/transformers/llama/test_modeling.py
@@ -101,7 +101,7 @@ class LlamaModelTester:
         self.return_dict = return_dict
 
     def prepare_config_and_inputs(self):
-        input_ids = ids_tensor([self.batch_size, self.seq_length], self.vocab_size)
+        input_ids = ids_tensor([self.batch_size, self.seq_length], self.vocab_size, dtype=paddle.int64)
 
         input_mask = None
         if self.use_input_mask:
